Split web socket server into separate module
